+++ src/main/java/dinkplugin/notifiers/data/LootNotificationData.java
@@ -10,6 +10,7 @@ import net.runelite.http.api.loottracker.LootRecordType;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 @Value
@@ -25,9 +26,12 @@ public class LootNotificationData extends NotificationData {
     @Nullable
     Double rarestProbability;
 
+    @Nullable
+    Collection<String> party;
+
     @Override
     public List<Field> getFields() {
-        List<Field> fields = new ArrayList<>(3);
+        List<Field> fields = new ArrayList<>(4);
         if (killCount != null) {
             fields.add(
                 new Field(
@@ -44,6 +48,9 @@ public class LootNotificationData extends NotificationData {
         );
         if (rarestProbability != null) {
             fields.add(new Field("Item Rarity", Field.formatProbability(rarestProbability)));
+        }
+        if (party != null && !party.isEmpty()) {
+            fields.add(new Field("Party Size", Field.formatBlock("", String.valueOf(party.size()))));
         }
         return fields;
     }